noobcash/requirements.txt
asgiref
certifi
chardet
click
Django
django-crispy-forms
django-static-jquery
Flask
Flask-Cors
idna
itsdangerous
Jinja2
MarkupSafe
prompt-toolkit
pycryptodome
Pygments
PyInquirer
pytz
regex
requests
six
sqlparse
texttable
urllib3
wcwidth
Werkzeug
wincertstore
